requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pymongo==4.6.0
pandas==2.1.4
scikit-learn==1.3.2
openai==1.3.7
exa-py>=2.0.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
pytest-mock==3.15.1
mongomock==4.1.2
python-multipart==0.0.18
email-validator==2.3.0
requests==2.31.0
